Unit test failures have been rectified. Integration test failures are expected, seeing as they each use minimization that is now lbfgs.

Old repository SHA1: c7e31c28628af3fc1104e937406a7b6636c78f81

+++ source/test/protocols/rotamer_recovery/RotamerRecovery.cxxtest.hh
@@ -162,7 +162,7 @@ public:
 
 			rr->run( pose_1ten_,*score_function_,*packer_task_1ten_);
 			rr->show( TR );
-			TS_ASSERT_DELTA( rr->recovery_rate(), Real(57)/Real(89) , .001 );
+			TS_ASSERT_DELTA( rr->recovery_rate(), Real(58)/Real(89) , .001 );
 		}
 
 		{
@@ -171,7 +171,7 @@ public:
 				"RRProtocolRTMin", "RRComparerAutomorphicRMSD", "RRReporterSimple"));
 			rr->run( pose_1ten_,*score_function_,*packer_task_1ten_);
 			rr->show( TR );
-			TS_ASSERT_DELTA( rr->recovery_rate(), Real(12)/Real(89) , .001 );
+			TS_ASSERT_DELTA( rr->recovery_rate(), Real(11)/Real(89) , .001 );
 		}
